Further changes

+++ shader.vert
@@ -1,5 +1,5 @@
 #version 450 core
-layout(location=0) in vec2 position;
+layout(location=0) in dvec2 position;
 layout(location=1) in vec4 color;
 
 uniform mat2 proj_matrix;
